Split reduction tests from test_math.py
